+++ packages/amplify-sdk/src/stores/token-store.js
@@ -4,7 +4,7 @@ import E from '../errors';
 import pluralize from 'pluralize';
 import snooplogg from 'snooplogg';
 
-const { log } = snooplogg('amplify-auth:token-store');
+const { log } = snooplogg('amplify-sdk:auth:token-store');
 const { highlight } = snooplogg.styles;
 
 /**