+++ packages/amplify-sdk/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.1.0
+
+ * fix(token-store): Updated `keytar` from v7.6.0 to v7.7.0, which now uses N-API. The Amplify SDK
+   no longer needs to install `keytar` at runtime which solves npm 7 related issues.
+ * fix(login): Validate token auth code before redirecting the browser to select the organization.
+ * chore: Updated dependencies.
+
 # v2.0.0 (Apr 21, 2021)
 
  * BREAKING CHANGE: Removed MBS and Titanium app development APIs.